// web/apps/webapp/src/queries/__generated__/teamPeopleBookings_addBookingMutation.graphql.ts
export type AddBookingInput = {
  clientMutationId?: string | null | undefined;
  customerId: string;
  deskIds: ReadonlyArray<string>;
  from: any;
  id?: string | null | undefined;
  locationId?: string | null | undefined;
  notes?: string | null | undefined;
  organizationId?: string | null | undefined;
  teamId?: string | null | undefined;
  to: any;
};
export type teamPeopleBookings_addBookingMutation$variables = {
  input: AddBookingInput;
};
export type teamPeopleBookings_addBookingMutation$data = {
  readonly addBooking: {
    readonly booking: {
      readonly customer: {
        readonly familyName: string | null | undefined;
        readonly givenName: string | null | undefined;
        readonly middleName: string | null | undefined;
        readonly name: string | null | undefined;
      };
      readonly desks: ReadonlyArray<{
        readonly locationTags: ReadonlyArray<{
          readonly name: string;
          readonly tagType: string | null | undefined;
          readonly uniqueId: string;
        }>;
        readonly name: string;
      }>;
      readonly from: any;
      readonly id: string;
      readonly location: {
        readonly name: string;
      } | null | undefined;
    };
  } | null | undefined;
};
export type teamPeopleBookings_addBookingMutation = {
  response: teamPeopleBookings_addBookingMutation$data;
  variables: teamPeopleBookings_addBookingMutation$variables;
};