// web/apps/webapp/src/queries/__generated__/teamPeopleBookings_allBookings_query.graphql.ts
export type teamPeopleBookings_allBookings_query$data = {
  readonly allBookings: ReadonlyArray<{
    readonly customer: {
      readonly familyName: string | null | undefined;
      readonly givenName: string | null | undefined;
      readonly middleName: string | null | undefined;
      readonly name: string | null | undefined;
      readonly photoUrl: string | null | undefined;
      readonly uniqueId: string;
    };
    readonly desks: ReadonlyArray<{
      readonly locationTags: ReadonlyArray<{
        readonly name: string;
        readonly tagType: string | null | undefined;
        readonly uniqueId: string;
      }>;
      readonly name: string;
    }>;
    readonly from: any;
    readonly id: string;
    readonly location: {
      readonly name: string;
    } | null | undefined;
    readonly team: {
      readonly name: string;
    } | null | undefined;
    readonly to: any;
  }> | null | undefined;
  readonly " $fragmentType": "teamPeopleBookings_allBookings_query";
};
export type teamPeopleBookings_allBookings_query$key = {
  readonly " $data"?: teamPeopleBookings_allBookings_query$data;
  readonly " $fragmentSpreads": FragmentRefs<"teamPeopleBookings_allBookings_query">;
};